+++ klodder/Classes/View_ColorPicker_OpacityGauge.mm
@@ -74,7 +74,7 @@
 		Rgba color = delegate.colorPickerState->Color_get();
 		CGColorSpaceRef colorSpace = CGColorSpaceCreateDeviceRGB();
 		CGFloat components[] = { color.rgb[0], color.rgb[1], color.rgb[2], 0.0f, color.rgb[0], color.rgb[1], color.rgb[2], 1.0f };
-		float locations[] = { 0.0f, 1.0f };
+		CGFloat locations[] = { 0.0f, 1.0f };
 		CGGradientRef gradient = CGGradientCreateWithColorComponents(colorSpace, components, locations, 2);
 		CGContextDrawLinearGradient(ctx, gradient, CGPointMake(0.0f, y), CGPointMake(self.frame.size.width, height), 0);
 		CGGradientRelease(gradient);
